+++ Rebus.MsgPack.Tests/MsgPackSerializerTests.cs
@@ -1,0 +1,8 @@
+﻿using NUnit.Framework;
+using Rebus.Tests.Serialization;
+
+namespace Rebus.MsgPack.Tests
+{
+    [TestFixture]
+    public class MsgPackSerializerTests : BasicSerializationTests<MsgPackSerializerFactory> { }
+}